requirements.txt
aiocache==0.12.2
aiohttp==3.9.3
alembic==1.13.1
anthropic==0.30.1
asyncpg==0.29.0
fastapi==0.109.2
gunicorn==21.2.0
httpx==0.25.2
json5==0.9.25
numpy==1.26.4
openai==1.35.10
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic[email]==2.7.0
python-dotenv==1.0.0
redis==5.0.7
requests==2.31.0
rq==1.16.2
sqlalchemy[asyncio]==2.0.25
uvicorn==0.27.0.post1
pytest==6.2.5
pytest-asyncio==0.18.3